requirements.txt
pyrogram==2.0.30
python-dotenv==0.21.0
smtplib